Moved Nether goo to the appropriate package (+ more annotations)
