+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/multiblocks/MultiBlockMachine.java
@@ -5,6 +5,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
 import org.apache.commons.lang.Validate;
 import org.bukkit.Material;
@@ -77,6 +78,7 @@ public abstract class MultiBlockMachine extends SlimefunItem implements NotPlace
         return displayRecipes;
     }
 
+    @Nonnull
     public MultiBlock getMultiBlock() {
         return multiblock;
     }
@@ -121,7 +123,9 @@ public abstract class MultiBlockMachine extends SlimefunItem implements NotPlace
 
                 return true;
             }
-            else return false;
+            else {
+                return false;
+            }
         };
     }
 
@@ -158,7 +162,8 @@ public abstract class MultiBlockMachine extends SlimefunItem implements NotPlace
         }
     }
 
-    protected Inventory findOutputChest(Block b, ItemStack output) {
+    @Nullable
+    protected Inventory findOutputChest(@Nonnull Block b, @Nonnull ItemStack output) {
         for (BlockFace face : outputFaces) {
             Block potentialOutput = b.getRelative(face);
 
@@ -183,6 +188,7 @@ public abstract class MultiBlockMachine extends SlimefunItem implements NotPlace
         return null;
     }
 
+    @Nonnull
     private static Material[] convertItemStacksToMaterial(ItemStack[] items) {
         List<Material> materials = new ArrayList<>();
