+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/guide/RecipeChoiceTask.java
@@ -3,6 +3,8 @@ package io.github.thebusybiscuit.slimefun4.implementation.guide;
 import java.util.HashMap;
 import java.util.Map;
 
+import javax.annotation.Nonnull;
+
 import org.apache.commons.lang.Validate;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -39,18 +41,18 @@ class RecipeChoiceTask implements Runnable {
      * @param inv
      *            The {@link Inventory} to start this task for
      */
-    public void start(Inventory inv) {
+    public void start(@Nonnull Inventory inv) {
         Validate.notNull(inv, "Inventory must not be null");
         inventory = inv;
         id = Bukkit.getScheduler().runTaskTimerAsynchronously(SlimefunPlugin.instance(), this, 0, UPDATE_INTERVAL).getTaskId();
     }
 
-    public void add(int slot, MaterialChoice choice) {
+    public void add(int slot, @Nonnull MaterialChoice choice) {
         Validate.notNull(choice, "Cannot add a null RecipeChoice");
         iterators.put(slot, new LoopIterator<>(choice.getChoices()));
     }
 
-    public void add(int slot, Tag<Material> tag) {
+    public void add(int slot, @Nonnull Tag<Material> tag) {
         Validate.notNull(tag, "Cannot add a null Tag");
         iterators.put(slot, new LoopIterator<>(tag.getValues()));
     }
