+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/resources/SlimefunResource.java
@@ -1,5 +1,8 @@
 package io.github.thebusybiscuit.slimefun4.implementation.resources;
 
+import javax.annotation.Nonnull;
+import javax.annotation.ParametersAreNonnullByDefault;
+
 import org.apache.commons.lang.Validate;
 import org.bukkit.NamespacedKey;
 import org.bukkit.inventory.ItemStack;
@@ -15,6 +18,7 @@ abstract class SlimefunResource implements GEOResource {
     private final int maxDeviation;
     private final boolean geoMiner;
 
+    @ParametersAreNonnullByDefault
     SlimefunResource(String key, String defaultName, ItemStack item, int maxDeviation, boolean geoMiner) {
         Validate.notNull(key, "NamespacedKey cannot be null!");
         Validate.notNull(defaultName, "The default name cannot be null!");
@@ -28,16 +32,19 @@ abstract class SlimefunResource implements GEOResource {
     }
 
     @Override
+    @Nonnull
     public NamespacedKey getKey() {
         return key;
     }
 
     @Override
+    @Nonnull
     public String getName() {
         return defaultName;
     }
 
     @Override
+    @Nonnull
     public ItemStack getItem() {
         return item.clone();
     }
