+++ src/main/java/me/mrCookieSlime/Slimefun/api/SlimefunItemStack.java
@@ -8,6 +8,9 @@ import java.util.Locale;
 import java.util.Optional;
 import java.util.function.Consumer;
 
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
 import org.apache.commons.lang.Validate;
 import org.bukkit.ChatColor;
 import org.bukkit.Color;
@@ -146,7 +149,7 @@ public class SlimefunItemStack extends CustomItem {
         setItemId(id);
     }
 
-    private void setItemId(String id) {
+    private void setItemId(@Nonnull String id) {
         Validate.notNull(id, "The Item id must never be null!");
         Validate.isTrue(id.equals(id.toUpperCase(Locale.ROOT)), "Slimefun Item Ids must be uppercase! (e.g. 'MY_ITEM_ID')");
 
@@ -169,7 +172,8 @@ public class SlimefunItemStack extends CustomItem {
      * 
      * @return The {@link SlimefunItem} id for this {@link SlimefunItemStack}
      */
-    public String getItemId() {
+    @Nonnull
+    public final String getItemId() {
         return id;
     }
 
@@ -179,6 +183,7 @@ public class SlimefunItemStack extends CustomItem {
      *
      * @return The {@link SlimefunItem} for this {@link SlimefunItemStack}, null if not found.
      */
+    @Nullable
     public SlimefunItem getItem() {
         return SlimefunItem.getByID(id);
     }
@@ -267,7 +272,11 @@ public class SlimefunItemStack extends CustomItem {
         return SkullItem.fromBase64(getTexture(id, texture));
     }
 
-    private static String getTexture(String id, String texture) {
+    @Nonnull
+    private static String getTexture(@Nonnull String id, @Nonnull String texture) {
+        Validate.notNull(id, "The id cannot be null");
+        Validate.notNull(texture, "The texture cannot be null");
+        
         if (texture.startsWith("ey")) {
             return texture;
         }
